Add a `SubscriptionRequest` test
