+++ client-js/test/client/subscription-request-test.js
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2019, TeamDev. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+
+import {SubscriptionRequest} from "@lib/client/client-request";
+import {Duration} from "@lib/client/time-utils";
+import {filteringRequestTest} from "./filtering-request-test";
+
+describe('SubscriptionRequest', function () {
+
+  const timeoutDuration = new Duration({seconds: 5});
+  this.timeout(timeoutDuration.inMs());
+
+  filteringRequestTest(newSubscriptionRequest, buildTopic, getTarget, getFieldMask);
+
+  function newSubscriptionRequest(targetType, clientStub, actorRequestFactory) {
+    return new SubscriptionRequest(targetType, clientStub, actorRequestFactory);
+  }
+
+  function buildTopic(request) {
+    return request.topic();
+  }
+
+  function getTarget(topic) {
+    return topic.getTarget();
+  }
+
+  function getFieldMask(topic) {
+    return topic.getFieldMask();
+  }
+});
